+++ app/src/main/java/com/fibelatti/photowidget/home/HomeViewModel.kt
@@ -67,7 +67,7 @@ class HomeViewModel @Inject constructor(
             photoWidgetStorage.getKnownWidgetIds(),
             photoWidgetStorage.getDraftWidgetIds(),
             updateSignal.receiveAsFlow(),
-        ) { ids: List<Int>, draftIds: List<Int>, _: Unit -> ids + draftIds }
+        ) { ids: List<Int>, draftIds: List<Int>, _: Unit -> ids + draftIds.filterNot(PhotoWidget::isDraftWidgetId) }
             .flatMapLatest { allIds: List<Int> ->
                 if (allIds.isEmpty()) return@flatMapLatest flowOf(emptyMap())