+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Fix "view previous" not working as expected when shuffle is enabled
 * Fix draft widgets not showing in "My Widgets"
+* Fix a crash that could happen when opening the photo picker
+* Fix a crash that could happen when the system tries to update the widget
 
 ## [v1.41.0] - 2026-04-18
 [v1.41.0]: https://github.com/fibelatti/photo-widget/releases/tag/v1.41.0